+++ lib/cleanup.js
@@ -34,6 +34,7 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 Object.defineProperty(exports, "__esModule", { value: true });
 const core = __importStar(require("@actions/core"));
 const utils_1 = require("./utils");
+const jobsummary_utils_1 = require("./jobsummary-utils");
 function cleanup() {
     return __awaiter(this, void 0, void 0, function* () {
         if (yield shouldSkipCleanup()) {
@@ -66,7 +67,7 @@ function cleanup() {
 function buildInfoPostTasks() {
     return __awaiter(this, void 0, void 0, function* () {
         const disableAutoBuildPublish = core.getBooleanInput(utils_1.Utils.AUTO_BUILD_PUBLISH_DISABLE);
-        const disableJobSummary = core.getBooleanInput(utils_1.Utils.JOB_SUMMARY_DISABLE) || !utils_1.Utils.isJobSummarySupported();
+        const disableJobSummary = core.getBooleanInput(utils_1.Utils.JOB_SUMMARY_DISABLE) || !jobsummary_utils_1.jobSummaryUtils.isJobSummarySupported();
         if (disableAutoBuildPublish && disableJobSummary) {
             core.info(`Both auto-build-publish and job-summary are disabled. Skipping Build Info post tasks.`);
             return;
@@ -94,11 +95,11 @@ function buildInfoPostTasks() {
 function hasUnpublishedModules(workingDirectory) {
     return __awaiter(this, void 0, void 0, function* () {
         // Save the old value of the environment variable to revert it later
-        const origValue = process.env[utils_1.Utils.JFROG_CLI_COMMAND_SUMMARY_OUTPUT_DIR_ENV];
+        const origValue = process.env[jobsummary_utils_1.jobSummaryUtils.JFROG_CLI_COMMAND_SUMMARY_OUTPUT_DIR_ENV];
         try {
             core.startGroup('Check for unpublished modules');
             // Avoid saving a command summary for this dry-run command
-            core.exportVariable(utils_1.Utils.JFROG_CLI_COMMAND_SUMMARY_OUTPUT_DIR_ENV, '');
+            core.exportVariable(jobsummary_utils_1.jobSummaryUtils.JFROG_CLI_COMMAND_SUMMARY_OUTPUT_DIR_ENV, '');
             // Running build-publish command with a dry-run flag to check if there are any unpublished modules, 'silent' to avoid polluting the logs
             const responseStr = yield utils_1.Utils.runCliAndGetOutput(['rt', 'build-publish', '--dry-run'], { cwd: workingDirectory });
             // Parse the JSON string to an object
@@ -111,7 +112,7 @@ function hasUnpublishedModules(workingDirectory) {
             return false;
         }
         finally {
-            core.exportVariable(utils_1.Utils.JFROG_CLI_COMMAND_SUMMARY_OUTPUT_DIR_ENV, origValue);
+            core.exportVariable(jobsummary_utils_1.jobSummaryUtils.JFROG_CLI_COMMAND_SUMMARY_OUTPUT_DIR_ENV, origValue);
             core.endGroup();
         }
     });
@@ -184,10 +185,10 @@ function generateJobSummary() {
         try {
             core.startGroup('Generating Job Summary');
             yield utils_1.Utils.runCli(['generate-summary-markdown']);
-            yield utils_1.Utils.setMarkdownAsJobSummary();
-            yield utils_1.Utils.populateCodeScanningTab();
+            yield jobsummary_utils_1.jobSummaryUtils.setMarkdownAsJobSummary();
+            yield jobsummary_utils_1.jobSummaryUtils.populateCodeScanningTab();
             // Clear files
-            yield utils_1.Utils.clearCommandSummaryDir();
+            yield jobsummary_utils_1.jobSummaryUtils.clearCommandSummaryDir();
         }
         catch (error) {
             core.warning('Failed while attempting to generate job summary: ' + error);
